requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
autoflake==2.3.0
certifi==2024.2.2
cffi==1.16.0
cryptography==41.0.7
defusedxml==0.7.1
diff-match-patch==20230430
distro==1.9.0
Django
django-cors-headers==4.3.1
django-import-export==3.3.7
django-rest-framework==0.1.0
djangorestframework==3.15.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
install==1.3.5
MarkupPy==1.14
odfpy==1.4.1
openai==1.27.0
openpyxl==3.1.2
passlib==1.7.4
pillow==10.2.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pyflakes==3.2.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
sniffio==1.3.1
sqlparse==0.4.4
tablib==3.5.0
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.10.0
xlrd==2.0.1
xlwt==1.3.0
requests